refactor gamehistory
